refactor: minify canvas code
