+++ packages/vrender/src/canvas/contributions/feishu/modules.ts
@@ -1,19 +1,5 @@
-import { ContainerModule } from 'inversify';
 import { FeishuCanvas } from './canvas';
 import { FeishuContext2d } from './context';
-import { CanvasFactory, Context2dFactory } from '../../constants';
-import type { CanvasConfigType, ICanvas } from '../../../interface';
+import { createModule } from '../create-canvas-module';
 
-export default new ContainerModule(bind => {
-  bind(CanvasFactory)
-    .toDynamicValue(() => {
-      return (params: CanvasConfigType) => new FeishuCanvas(params);
-    })
-    .whenTargetNamed(FeishuCanvas.env);
-
-  bind(Context2dFactory)
-    .toDynamicValue(() => {
-      return (params: ICanvas, dpr: number) => new FeishuContext2d(params, dpr);
-    })
-    .whenTargetNamed(FeishuContext2d.env);
-});
+export default createModule(FeishuCanvas, FeishuContext2d);
