+++ packages/vrender/src/canvas/contributions/lynx/modules.ts
@@ -1,19 +1,5 @@
-import { ContainerModule } from 'inversify';
 import { LynxCanvas } from './canvas';
 import { LynxContext2d } from './context';
-import { CanvasFactory, Context2dFactory } from '../../constants';
-import type { CanvasConfigType, ICanvas } from '../../../interface';
+import { createModule } from '../create-canvas-module';
 
-export default new ContainerModule(bind => {
-  bind(CanvasFactory)
-    .toDynamicValue(() => {
-      return (params: CanvasConfigType) => new LynxCanvas(params);
-    })
-    .whenTargetNamed(LynxCanvas.env);
-
-  bind(Context2dFactory)
-    .toDynamicValue(() => {
-      return (params: ICanvas, dpr: number) => new LynxContext2d(params, dpr);
-    })
-    .whenTargetNamed(LynxContext2d.env);
-});
+export default createModule(LynxCanvas, LynxContext2d);
