+++ packages/vrender/src/canvas/contributions/tt/modules.ts
@@ -1,19 +1,5 @@
-import { ContainerModule } from 'inversify';
 import { TTCanvas } from './canvas';
 import { TTContext2d } from './context';
-import { CanvasFactory, Context2dFactory } from '../../constants';
-import type { CanvasConfigType, ICanvas } from '../../../interface';
+import { createModule } from '../create-canvas-module';
 
-export default new ContainerModule(bind => {
-  bind(CanvasFactory)
-    .toDynamicValue(() => {
-      return (params: CanvasConfigType) => new TTCanvas(params);
-    })
-    .whenTargetNamed(TTCanvas.env);
-
-  bind(Context2dFactory)
-    .toDynamicValue(() => {
-      return (params: ICanvas, dpr: number) => new TTContext2d(params, dpr);
-    })
-    .whenTargetNamed(TTContext2d.env);
-});
+export default createModule(TTCanvas, TTContext2d);
